bot/requirements.txt
telebot
numpy
pillow
requests